requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# ai-tts-donations/bot.py: 16
python_dotenv == 0.19.0

# ai-tts-donations/bot.py: 18
requests == 2.25.1

# ai-tts-donations/bot.py: 11,12,13,17
twitchAPI == 2.4.2
